+++ packages/react-core/src/components/Drawer/__tests__/Drawer.test.tsx
@@ -1,12 +1,13 @@
 import {
   Drawer,
-  DrawerPanelContent,
-  DrawerContent,
-  DrawerHead,
   DrawerActions,
   DrawerCloseButton,
+  DrawerContent,
   DrawerContentBody,
-  DrawerPanelBody
+  DrawerHead,
+  DrawerPanelBody,
+  DrawerColorVariant,
+  DrawerPanelContent
 } from '../';
 import React from 'react';
 import { mount } from 'enzyme';
@@ -68,9 +69,9 @@ test(`Drawer expands from bottom`, () => {
   expect(view).toMatchSnapshot();
 });
 
-test(`Drawer has resizable css`, () => {
+test(`Drawer has resizable css and color variants`, () => {
   const panelContent = (
-    <DrawerPanelContent isResizable minSize={'200px'} defaultSize={'300px'} maxSize={'400px'}>
+    <DrawerPanelContent isResizable minSize={'200px'} defaultSize={'300px'} maxSize={'400px'} colorVariant={DrawerColorVariant.light200}>
       <DrawerHead>
         <span>drawer-panel</span>
         <DrawerActions>