+++ packages/react-code-editor/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.32](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-code-editor@4.1.31...@patternfly/react-code-editor@4.1.32) (2021-03-04)
+
+**Note:** Version bump only for package @patternfly/react-code-editor
+
+
+
+
+
+## [4.1.31](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-code-editor@4.1.30...@patternfly/react-code-editor@4.1.31) (2021-03-03)
+
+**Note:** Version bump only for package @patternfly/react-code-editor
+
+
+
+
+
+## [4.1.30](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-code-editor@4.1.29...@patternfly/react-code-editor@4.1.30) (2021-03-03)
+
+**Note:** Version bump only for package @patternfly/react-code-editor
+
+
+
+
+
+## [4.1.29](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-code-editor@4.1.28...@patternfly/react-code-editor@4.1.29) (2021-03-02)
+
+**Note:** Version bump only for package @patternfly/react-code-editor
+
+
+
+
+
 ## [4.1.28](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-code-editor@4.1.27...@patternfly/react-code-editor@4.1.28) (2021-03-02)
 
 **Note:** Version bump only for package @patternfly/react-code-editor