+++ packages/react-topology/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.41](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.7.40...@patternfly/react-topology@4.7.41) (2021-03-04)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [4.7.40](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.7.39...@patternfly/react-topology@4.7.40) (2021-03-03)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [4.7.39](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.7.38...@patternfly/react-topology@4.7.39) (2021-03-03)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [4.7.38](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.7.37...@patternfly/react-topology@4.7.38) (2021-03-02)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
 ## [4.7.37](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@4.7.36...@patternfly/react-topology@4.7.37) (2021-03-02)
 
 **Note:** Version bump only for package @patternfly/react-topology